move the lsb_fallback to the lsb module
