+++ ceph_deploy/lsb.py
@@ -1,5 +1,9 @@
-from . import hosts
+import logging
 from . import exc
+
+
+logger = logging.getLogger(__name__)
+
 
 def check_lsb_release():
     """
@@ -17,6 +21,23 @@ def check_lsb_release():
     ret = process.wait()
     if ret != 0:
         raise RuntimeError('The lsb_release command was not found on remote host.  Please install the lsb-release package.')
+
+
+def lsb_fallback(conn):
+    """
+    This fallback will attempt to detect the distro, release and codename for
+    a given remote host when lsb fails. It uses the
+    ``platform.linux_distribution`` module that should be fairly robust and
+    would prevent us from adding repositories and installing a package just to
+    detect a platform.
+    """
+    distro, release, codename = conn.modules.platform.linux_distribution()
+    return (
+        str(distro).rstrip(),
+        str(release).rstrip(),
+        str(codename).rstrip()
+    )
+
 
 def lsb_release():
     """
@@ -82,7 +103,8 @@ def get_lsb_release(sudo):
         check_lsb_release_r = sudo.compile(check_lsb_release)
         status = check_lsb_release_r()
     except RuntimeError as e:
-        return hosts.lsb_fallback(sudo)
+        logger.warning('lsb_release was not found - inferring OS details')
+        return lsb_fallback(sudo)
 
     lsb_release_r = sudo.compile(lsb_release)
     return lsb_release_r()
